Add support for configuration overrides.
Support for overriding file upload preferences was added to
ExpressionEngine in version 2.4.0. The overrides are applied when
accessing the upload preferences through the file upload preferences
model. This commit modifies file randomization to use the file upload
preferences model to retrieve the server_path instead of accessing the
database directly.